test(core): split real library tests
